rename: 탭 별 컴포넌트 폴더 구분 및 이동
