+++ src/app/property/[id]/components/detail/PropertyDetailTab.tsx
