+++ src/app/property/[id]/components/detail/PropertyInfoBox.tsx
