+++ src/app/property/[id]/components/detail/ScrollableTabBar.tsx
